stack

+++ dataAlgorithm/chenmingming/stack/stack_testMain.cpp
@@ -5,7 +5,7 @@ using namespace std;
 int main()
 {
     int nums = 0;
-    for(size_t i = 8; i < 10; ++i)
+    for(int i = 8; i < 10; ++i)
     {
         Stack<int> intstack;
         if(intstack.Empty())
